Issue #615 - further fixes for issues in free5gc e2e suite
- adjustments to usage of Kubernetes IO annotation
  "internal.config.kubernetes.io/path" to keep it present
  consistently and avoid it showing up as a diff in update
  operations
- CLI E2E tests now respect a CLEANUP_ON_FAIL variable -
  if run with CLEANUP_ON_FAIL='false' and a CLI test fails,
  the suite will not clean that test's namespace off the Kind
  cluster, allowing for easier inspection of the state at
  at time of failure

https://github.com/nephio-project/nephio/issues/615

+++ test/e2e/cli/suite.go
@@ -117,9 +117,13 @@ func (s *CliTestSuite) RunTestCase(t *testing.T, tc TestCaseConfig) {
 	repoURL := s.GitServerURL + "/" + testGitUserOrg + "/" + strings.ReplaceAll(tc.TestCase, "/", "-")
 
 	KubectlCreateNamespace(t, tc.TestCase)
+
+	failed := false
 	t.Cleanup(func() {
-		KubectlDeleteNamespace(t, tc.TestCase)
-		deleteRemoteTestRepo(t, tc.TestCase)
+		if !(failed && os.Getenv("CLEANUP_ON_FAIL") == "false") {
+			KubectlDeleteNamespace(t, tc.TestCase)
+			deleteRemoteTestRepo(t, tc.TestCase)
+		}
 	})
 
 	createRemoteTestRepo(t, tc.TestCase)
@@ -201,10 +205,7 @@ func (s *CliTestSuite) RunTestCase(t *testing.T, tc TestCaseConfig) {
 		}
 
 		if failureOutput != "" {
-			t.Fatalf("\n%s\n%s", getTestFailureMessage(i, tc), failureOutput)
-		}
-
-		if failureOutput != "" {
+			failed = true
 			t.Fatalf("\n%s\n%s", getTestFailureMessage(i, tc), failureOutput)
 		}
